db_research/requirements.txt
beanie==1.26.0
motor~=3.5.1
pydantic-settings==2.3.4
pydantic==2.8.2
fastapi~=0.111.1
uvicorn~=0.30.1
pymongo~=4.8.0
Faker~=26.0.0
asyncpg==0.29.0
python-dotenv~=1.0.1